% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{as_marginal}
\alias{as_marginal}
\title{Convert a DGM to the marginal parameterization}
\usage{
as_marginal(dgm, times)
}
\arguments{
\item{dgm}{A \code{dgm}.}

\item{times}{Grid on which to evaluate.}
}
\value{
A \code{dgm_marginal}.
}
\description{
Always exact: \verb{V = Z G Z' + sigma^2 I}.
}
