% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{cov_at}
\alias{cov_at}
\title{Covariance of a data-generating mechanism at given times}
\usage{
cov_at(dgm, times)
}
\arguments{
\item{dgm}{A \code{dgm} object.}

\item{times}{Numeric vector of observation times.}
}
\value{
A covariance matrix of dimension \code{length(times)}.
}
\description{
The single polymorphic operation both parameterizations answer.
}
