% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{null_fit}
\alias{null_fit}
\title{A failed fit}
\usage{
null_fit(converged = FALSE)
}
\arguments{
\item{converged}{Logical, normally \code{FALSE}.}
}
\value{
A \code{fit_result} of \code{NA}s.
}
\description{
A failed fit
}
